# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maodo Services API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, processing, completed, cancelled]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/ref/{reference}:
    get:
      operationId: getOrderByReference
      tags: [orders]
      summary: Get an order by reference code
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/orders:
    delete:
      operationId: clearOrderHistory
      tags: [orders]
      summary: Delete all orders matching an optional status filter (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, processing, completed, cancelled]
      responses:
        "200":
          description: Orders deleted
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: integer
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}:
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete a specific order (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order deleted
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: integer
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    OrderInput:
      type: object
      required: [serviceId, clientName, clientPhone, description]
      properties:
        serviceId:
          type: string
        clientName:
          type: string
          minLength: 2
        clientPhone:
          type: string
          minLength: 6
        clientEmail:
          type: ["string", "null"]
        description:
          type: string
          minLength: 10
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
    Order:
      type: object
      required: [id, reference, serviceId, serviceName, servicePrice, clientName, clientPhone, description, status, paymentStatus, createdAt, updatedAt]
      properties:
        id:
          type: integer
        reference:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        servicePrice:
          type: integer
        serviceDelay:
          type: string
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: ["string", "null"]
        description:
          type: string
        fileName:
          type: ["string", "null"]
        filePath:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        paymentReference:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        transactionId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...